+++ Shape/Shape/RectangleFactorySingleton.cpp
@@ -2,7 +2,7 @@
 #include "RectangleFactorySingleton.h"
 #include "Rectangle.h"
 
-std::unique_ptr<AbstractShape> RectangleFactorySingleton::CreateShape(double a, double b)
+std::unique_ptr<AbstractShape> RectangleFactorySingleton::CreateShape(BigInt a, BigInt b)
 {
     return std::make_unique<CRectangle>(a, b);
 }
