+++ Shape/Shape/TriangleFactorySingleton.cpp
@@ -2,7 +2,7 @@
 #include "TriangleFactorySingleton.h"
 #include "Triangle.h"
 
-std::unique_ptr<AbstractShape> TriangleFactorySingleton::CreateShape(double a, double b, double c)
+std::unique_ptr<AbstractShape> TriangleFactorySingleton::CreateShape(BigInt a, BigInt b, BigInt c)
 {
     return std::make_unique<CTriangle>(a, b, c);
 }
